+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v1.3.6 - Requires SelfHelp v7.0.0+ (Not released)
 ### New Features
  - add computability with `user_input` refactoring from SelfHelp v7.0.0 
+ - when a new survey is created it is automatically added to the dataTables.
+ - when a survey `title` is set, it is used for `displayName` for the `dataTables`
 ### Bugfix
  - check if there is a selected survey in the style before trying to save
  - do not try to save the survey in CMS mode